refactored weathergrabber application package
