+++ weathergrabber/adapter/tty/console_tty.py
@@ -2,7 +2,7 @@ from weathergrabber.application.usecases.use_case import UseCase
 from weathergrabber.domain.adapter.params import Params
 from weathergrabber.domain.adapter.icon_enum import IconEnum
 from weathergrabber.domain.entities.weather_icon_enum import WeatherIconEnum
-from weathergrabber.weathergrabber_application import WeatherGrabberApplication
+from weathergrabber.application.weathergrabber_application import WeatherGrabberApplication
 import logging
 
 class ConsoleTTY:
